Scripts/requirement.txt
opencv-python
mediapipe
numpy
tensorflow
pillow
pyttsx3